Now basically supports 1920x1080 and 1280x720 screen sizes. Simply change window size in project settings

+++ src/core/ViewportCardFocus.gd
@@ -21,6 +21,11 @@ onready var viewport = $ViewportContainer/Viewport
 var canonical_size:= Vector2(0,0)
 var vbc_rect_offset:= Vector2(0,0)
 
+#I was having an impossible time getting the viewport/camera system to work
+#in 1280x720. Instead I'm using a different method there, adding the cards
+#directly into $VBC. It's gross but it works
+var vbc_position_mode = true
+
 # Called when the node enters the scene tree for the first time.
 func _ready():
 	cfc.map_node(self)
@@ -42,35 +47,80 @@ func _ready():
 		container.re_place()
 	focus_info.info_panel_scene = info_panel_scene
 	focus_info.setup()
+	
+	if vbc_position_mode:
+		$VBC.remove_child(card_focus)
+		$VBC.remove_child(focus_info)
 
-func reposition():
+func reposition_vbc():
 	var mouse_pos: Vector2
 	if gamepadHandler.is_mouse_input():
 		mouse_pos = get_global_mouse_position()	
 	else:
 		mouse_pos = gamepadHandler.get_approx_position()	
+	
+	$VBC.margin_right = 0
+	$VBC.margin_bottom = 0
+	
+	var display_size = Vector2(0,0)
+	var display_position = Vector2(0,0)
+	var viewport_size = get_viewport().size
+	var spacer = 15
+	if _current_focus_source and is_instance_valid(_current_focus_source):	
+		var card = _current_focus_source
+		var multiplier =  card.focused_scale * cfc.curr_scale		
+		var card_size = card.canonical_size * multiplier
+		if (card._horizontal and card.get_is_faceup()):
+			$VBC.rect_rotation = 90
+			vbc_rect_offset = Vector2 (card_size.y, 0)
+			display_size = Vector2(card_size.y, card_size.x)
+		else:
+			$VBC.rect_rotation = 0
+			vbc_rect_offset = Vector2 (0,0)
+			display_size = Vector2(card_size.x, card_size.y)
+		if mouse_pos.x + display_size.x >= viewport_size.x :
+			display_position = Vector2( spacer, spacer)
+		else:
+			display_position = Vector2(viewport_size.x - display_size.x - spacer,  spacer)
+		pass
+
+	$VBC.rect_position = display_position + vbc_rect_offset
+
+func reposition():
+	if vbc_position_mode:
+		reposition_vbc()
+		return
 		
+	var mouse_pos: Vector2
+	if gamepadHandler.is_mouse_input():
+		mouse_pos = get_global_mouse_position()	
+	else:
+		mouse_pos = gamepadHandler.get_approx_position()	
+	
+	var viewport_size = get_viewport().size
+			
 	if _current_focus_source and is_instance_valid(_current_focus_source)\
 			and _current_focus_source.get_state_exec() != "pile"\
 			and cfc.game_settings.focus_style == CFInt.FocusStyle.BOTH_INFO_PANELS_ONLY:
-		if mouse_pos.y + focus_info.rect_size.y/2 > get_viewport().size.y:
-			$VBC.rect_position.y = get_viewport().size.y - focus_info.rect_size.y
+		if mouse_pos.y + focus_info.rect_size.y/2 > viewport_size.y:
+			$VBC.rect_position.y = viewport_size.y - focus_info.rect_size.y
 		else:
 			$VBC.rect_position.y = mouse_pos.y - focus_info.rect_size.y / 2
-		if mouse_pos.x + focus_info.rect_size.x + 60 > get_viewport().size.x:
-			$VBC.rect_position.x = get_viewport().size.x - focus_info.rect_size.x
+		if mouse_pos.x + focus_info.rect_size.x + 60 > viewport_size.x:
+			$VBC.rect_position.x = viewport_size.x - focus_info.rect_size.x
 			$VBC.rect_position.y = mouse_pos.y - 500
 		else:
 			$VBC.rect_position.x = mouse_pos.x + 60
 
 	elif _current_focus_source and is_instance_valid(_current_focus_source)\
-			and mouse_pos.x > get_viewport().size.x - canonical_size.x*2.5\
+			and mouse_pos.x > viewport_size.x - canonical_size.x*2.5\
 			and mouse_pos.y < canonical_size.y*2:
 		$VBC.rect_position.x = 0
 		$VBC.rect_position.y = 0
 	elif _current_focus_source:
-		$VBC.rect_position.x = get_viewport().size.x - vbc_rect_offset.x
+		$VBC.rect_position.x = viewport_size.x - vbc_rect_offset.x
 		$VBC.rect_position.y = 0
+
 
 	if not is_instance_valid(_current_focus_source)\
 			and $VBC/Focus.modulate.a != 0\
@@ -92,9 +142,9 @@ func garbage_collection():
 			to_delete.append(c)
 		# We don't delete old dupes, to avoid overhead to the engine
 		# insteas, we just hide them.
-		if _current_focus_source != c\
-				and not $VBC/Focus/Tween.is_active():
-			current_dupe_focus.visible = false
+		if _current_focus_source != c:
+			if vbc_position_mode or not $VBC/Focus/Tween.is_active():
+				current_dupe_focus.visible = false
 	for c in to_delete:
 		var _found = _previously_focused_cards.erase(c)
 
@@ -128,7 +178,11 @@ func focus_card(card: Card, show_preview := true) -> void:
 			# We display a "pure" version of the card
 			# This means we hide buttons, tokens etc
 			dupe_focus.set_state(Card.CardState.VIEWPORT_FOCUS)
-			_focus_viewport.add_child(dupe_focus)
+			if vbc_position_mode:
+				dupe_focus.set_position(Vector2(0,0))						
+				$VBC.add_child(dupe_focus)
+			else:
+				_focus_viewport.add_child(dupe_focus)
 			_extra_dupe_ready(dupe_focus, card)
 			dupe_focus.is_faceup = card.is_faceup
 			dupe_focus.is_viewed = card.is_viewed
@@ -157,6 +211,11 @@ func focus_card(card: Card, show_preview := true) -> void:
 		# We store all our previously focused cards in an array, and clean them
 		# up when they're not focused anymore
 		_previously_focused_cards[card] = dupe_focus
+		
+		if vbc_position_mode:
+			garbage_collection()
+			reposition()
+			return
 		
 		set_camera_position(dupe_focus)
 
@@ -189,8 +248,8 @@ func focus_card(card: Card, show_preview := true) -> void:
 		else:
 			$VBC.rect_rotation = 0
 		
-	garbage_collection()
-	reposition()		
+		garbage_collection()
+		reposition()		
 
 func set_camera_position(dupe_focus):
 	if !_current_focus_source or !is_instance_valid(_current_focus_source):
@@ -200,7 +259,6 @@ func set_camera_position(dupe_focus):
 	# see https://github.com/godotengine/godot/issues/3393
 	# We make the viewport camera focus on it
 	_focus_camera.position = dupe_focus.global_position
-	
 	#horizontal case
 	if (_current_focus_source._horizontal) and _current_focus_source.get_state_exec() != "pile":
 		canonical_size = Vector2(_current_focus_source.canonical_size.y, _current_focus_source.canonical_size.x )	
@@ -208,33 +266,34 @@ func set_camera_position(dupe_focus):
 		_focus_camera.set_offset(Vector2(0, -20))
 	#normal case
 	else:
-		vbc_rect_offset = $VBC.rect_size
-		canonical_size = _current_focus_source.canonical_size
+		vbc_rect_offset =  $VBC.rect_size
+		canonical_size = _current_focus_source.canonical_size 
 		_focus_camera.set_offset(Vector2(0, 0))
 
 # Hides the focus viewport when we're done looking at it
 func unfocus(card: Card) -> void:
 	if _current_focus_source == card:
 		_current_focus_source = null
-		$VBC/Focus/Tween.remove_all()
-		$VBC/Focus/Tween.interpolate_property($VBC/Focus,'modulate',
-				$VBC/Focus.modulate, Color(1,1,1,0), 0.25,
-				Tween.TRANS_SINE, Tween.EASE_IN)
-		if focus_info.modulate != Color(1,1,1,0):
-			$VBC/Focus/Tween.interpolate_property(focus_info,'modulate',
-					focus_info.modulate, Color(1,1,1,0), 0.25,
+		if !vbc_position_mode:
+			$VBC/Focus/Tween.remove_all()
+			$VBC/Focus/Tween.interpolate_property($VBC/Focus,'modulate',
+					$VBC/Focus.modulate, Color(1,1,1,0), 0.25,
 					Tween.TRANS_SINE, Tween.EASE_IN)
-		$VBC/Focus/Tween.start()
-	garbage_collection()
-	reposition()
+			if focus_info.modulate != Color(1,1,1,0):
+				$VBC/Focus/Tween.interpolate_property(focus_info,'modulate',
+						focus_info.modulate, Color(1,1,1,0), 0.25,
+						Tween.TRANS_SINE, Tween.EASE_IN)
+			$VBC/Focus/Tween.start()
+		garbage_collection()
+		reposition()
 
 
 # Tells the currently focused card to stop focusing.
 func unfocus_all() -> void:
 	if _current_focus_source:
 		_current_focus_source.set_to_idle()
-	garbage_collection()
-	reposition()
+		garbage_collection()
+		reposition()
 
 
 # Overridable function for games to extend preprocessing of dupe card
@@ -250,13 +309,15 @@ func _extra_dupe_preparation(dupe_focus: Card, card: Card) -> void:
 # warning-ignore:unused_argument
 # warning-ignore:unused_argument
 func _extra_dupe_ready(dupe_focus: Card, card: Card) -> void:
+	var multiplier =  dupe_focus.focused_scale * cfc.curr_scale 
 	if CFConst.VIEWPORT_FOCUS_ZOOM_TYPE == "scale":
-		dupe_focus.scale = Vector2(1,1) * dupe_focus.focused_scale * cfc.curr_scale
+		dupe_focus.scale = Vector2(1,1) * multiplier
 	else:
-		dupe_focus.resize_recursively(dupe_focus._control, dupe_focus.focused_scale * cfc.curr_scale)
-		dupe_focus.get_card_front().scale_to(dupe_focus.focused_scale * cfc.curr_scale)
+		dupe_focus.resize_recursively(dupe_focus._control, multiplier)
+		dupe_focus.get_card_front().scale_to( multiplier)
 		dupe_focus.tokens.token_drawer(false, true)
 
+	dupe_focus.scale = Vector2(2,2) 
 func _input(event):
 	# We use this to allow the developer to take card screenshots
 	# for any number of purposes
@@ -282,7 +343,9 @@ func resize():
 		return
 
 	if is_instance_valid(get_viewport()):
+		$ViewportContainer.rect_min_size = get_viewport().size
 		$ViewportContainer.rect_size = get_viewport().size
+
 
 
 func toggle_glow(is_enabled := true) -> void:
